Adding the wonderful Parallel Task Extension Nuget Library to the project.
This gives me an observable Dictionary that I used for all the Properties found in all the projects. It needs to be observable, because as properties are removed, the UI (ListView) needs to stay updated.

By using this class, I was able to remove  a container class and simplify the code.

Very nice.