General Cleanup
I did a lot here:
I bound more ItemSources in XAML, instead of the code behind. to make it clearer where the data source was.

I also added some calls to OnPropertyChanged so subscribers can be informed of property changes in the view model.
This has broken drag and drop for now: But I will fix this later.

I also cleaned up the layout of alot of the UI Controls. So instead of using hard-coded positions, they more freely float. And hence respond better to resizing of the main window.
This includes a total overhaul of the global properties tab.

I further augmented the viewmodel to return properties from the backing model. And to respond better when the backing model has not been instantiated yet.