Synchronized property selection with the properties display.
Now when the user selects a different configuration or platform, the list of properties in the props window will update accordingly.
It will refresh the list and show all properties that are returned from the msbuild project that are evaluated for that condition of configuration and platform.
One last thing I had to do in order to get this to work was change the container holding the ReferenceProperties from a Dictionary to an Observable container. Otherwise it simply didn't work.

I also got rid of the event handler in the code behind, and used pure XAML to bind the details ListView with the Properties List view. I had to write my very first converter to do it (Trivially easy). It now works very well.