Finally got drag and drop to work for moving values to the property sheet.
Drag and Drop is only supported (for now) from the bottom listview. It is not supported from the top one.
ON the other hand, the top listview supports simply deleting a property and NOT moving it over to the property sheet.

This is way more elegant than my earlier versions from a few weeks ago. This makes classes operate on their own data, and not someone elses.