Refactoring the msbuild project files.
I added a master property sheet (Using this tool itself) to consolidate
all the build properties down to one file.

While doing so, I found that if I tried to delete the
  Platform
  Configuration
properties, then it would appear to let me delete them from the UI only.
Since they are global properties, it wouldn't really let me delete them.
So I have to fix that.

Now all the output is in one directory, and this solution is much more
organized.