Fixed a bug in the OutputPath property values not changing
Problem:
Given that a CSProject would cache it's value for the OutputPath property.
This would be cached for the configuration and platform from when the CSProject was instantiated. But if the configuration and platform changed,
then the OutputPath would not get updated: Because it was cached.
Thus it would return incorrect results.

Fixed by not caching the property value anymore.