New Feature: mass property remover
Given an existing property sheet. Take all the properties from it and remove them
from all the projects (For a given configuration).

This presents a completely different workflow than from what I originally
had envisoned from this application:
 Instead of dragging and dropping properties from right to left this
removes properties all properties from the right that are found in the left.
This is for a situation where the user has already written a valid property sheet
and simply wants to apply it to all the projects on hand.
Unfortunately the limitation here is that it stil works on only one build configuration
at a time. So you have to manually go and change each configuration, apply this
operation, save the files, and repeat.
I know of a quicker way to do this for all configurations, and that is to use a different
model that Microsoft offers: The Microsoft.Build.Construction XML model instead. And hence
simply not use the Evaluation model this application is built on.
But that will be for a different changelist and a different day.