Added mass property replace from a property sheet that works in all configurations
This one of the most powerful new features yet that I have put in.
Given a property sheet, you can say, given a property in a property sheet, find all instances
of that property in all projects for all configurations and platforms and delete them from the
project files. Leaving the property sheet to be the sole definition for that property.

Previously I had this working for one configuration/platform only. Now it works for all which
is really really cool.