Make Property Value handling case-insensitive
I have now made the display and internal handling of property values case insensitive.
This was one shortcut I took early on to speed development. But now I can afford to pay
off this debt.

The key was to get the source code for ObservableConcurrentDictionary and add an extra
constructor that takes an IEqualityComparer<> type.
Oh, and I had to remove all the calls to 'ToLower'

Once that was done, I was able to remove the reference to the NugetPackage for the ParalleExtensionExtras.