Fixed a bug related to loading two different directories
Problem:
1. Load projects in directory A.
2. Note the available configurations and platforms in the global properties list
3. Load projects in directory B.
3. Note the configurations and platforms in the global properties list have not changed.

Solution:
Needed to reset the configuration and platform list when it's populated.
It it has elements in it when the directory is loaded, it empties it out first.

I found this problem yesterday while demonstrating this tool to someone.

I also tested reloading two different property sheets, and found that worked just fine.