Fix file save code
Background:
The code that saved a file, checked if the property sheet was attached.
And if it was not attached, it would attempt to create an XML Import element
pointing to the property sheet.

Problem:
Unfortunately the property sheet would get attached over and over again if the
save button was pressed multiple times.

Analysis:
This is because the project was not getting ReEvaluated properly as I would expect.
Hence the Import would be saved on disk, but the logical Project instance would not
see it.

Solution:
Use XLinq to query the file, instead of the msbuild Project instance, to see if the import is attached already.