Remove property from all configurations
I have long wanted to be able to remove one property
from all configurations at once. Previously drag and drop
of a property to the property sheet would simply remove
the property from that one configuration only. That's great
for the one configuration, but when there are a dozen
configurations, it's a pain to switch between each one
and then select the particular value of the property you want
to remove, or move.

This change enables removing a property only from files
that contain that value. It will remove all instances of that
property for all configs in that file.

I also fixed it so that it marks the file as dirty.

I also fixed it so that a file save all, will only save dirty files.