Add feature to find projects that output to a particular build directory
This is an optional tool that allows the user to discover which projects
output to a particular build directory. It can be useful to help discover which
projects are 'dead' projects. Where a dead project is one that is simply
abandoned and does not build to the final output directory.